# Optimize prefix matching