--- a/pkg/storegateway/indexheader/header.go
+++ b/pkg/storegateway/indexheader/header.go
@@ -36,8 +36,9 @@ type Reader interface {
 	// LabelValues returns all label values for given label name or error.
 	// If no values are found for label name, or label name does not exists,
 	// then empty string is returned and no error.
+	// If non-empty prefix is provided, only values starting with the prefix are returned.
 	// If non-nil filter is provided, then only values for which filter returns true are returned.
-	LabelValues(name string, filter func(string) bool) ([]string, error)
+	LabelValues(name string, prefix string, filter func(string) bool) ([]string, error)
 
 	// LabelNames returns all label names in sorted order.
 	LabelNames() ([]string, error)