--- a/pkg/storegateway/indexheader/index/postings.go
+++ b/pkg/storegateway/indexheader/index/postings.go
@@ -386,10 +386,7 @@ func (t *PostingOffsetTableV2) LabelValues(name string, prefix string, filter fu
 	if prefix != "" {
 		// Find the first offset that is greater or equal to the prefix.
 		offsetIdx = sort.Search(len(e.offsets), func(i int) bool {
-			if prefix <= e.offsets[i].value {
-				return true
-			}
-			return false
+			return prefix <= e.offsets[i].value
 		})
 
 		// We always include the last value in the offsets,