… go.sum
… pkg/storegateway/indexheader/reader_benchmarks_test.go
… vendor/github.com/prometheus/prometheus/model/labels/matcher.go
… vendor/modules.txt