#### What this PR does

When a label has lots of values (for example `__name__` can have tens of thousands of values) and a matcher is looking for a prefix (for example, `__name__=~"keda_.*"`) we were listing all the values and applying the match function to all of them.

However, we know the offset of each 32nd label, so we know more or less where we can start searching for that prefix, and we know that label values are sorted lexicographically, so we know we should stop looking if we stop seeing that prefix.

This implements that optimization, and also adds benchmark cases for it (label `i` has 500k values, each one matching 50 series).

```
name                                                                        old time/op    new time/op    delta
BucketIndexReader_ExpandedPostings/n="X"-16                                   1.92µs ±13%    1.76µs ±20%   -8.53%  (p=0.036 n=10+20)
BucketIndexReader_ExpandedPostings/n="X",j="foo"-16                           2.69µs ±11%    2.77µs ±16%     ~     (p=0.328 n=10+20)
BucketIndexReader_ExpandedPostings/n="X",j!="foo"-16                          2.74µs ±13%    2.71µs ±11%     ~     (p=0.507 n=10+20)
BucketIndexReader_ExpandedPostings/j=~"XXX|YYY"-16                            1.98µs ±20%    1.97µs ± 9%     ~     (p=0.542 n=10+19)
BucketIndexReader_ExpandedPostings/j=~"X.+"-16                                2.74µs ±12%    1.83µs ± 8%  -33.10%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~"X|Y|Z"-16                              2.24µs ± 9%    2.19µs ±12%     ~     (p=0.282 n=10+20)
BucketIndexReader_ExpandedPostings/n="1"-16                                   8.00ms ±13%    7.62ms ±18%     ~     (p=0.091 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-16                           27.5ms ± 2%    27.7ms ± 0%   +0.57%  (p=0.040 n=10+17)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-16                           27.5ms ± 1%    27.7ms ± 1%   +0.86%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-16                          23.3ms ± 1%    23.6ms ± 1%   +1.52%  (p=0.000 n=9+19)
BucketIndexReader_ExpandedPostings/i=~".*"-16                                  107ms ± 4%     108ms ± 3%     ~     (p=0.091 n=10+20)
BucketIndexReader_ExpandedPostings/i=~".+"-16                                  295ms ± 2%     299ms ± 3%   +1.41%  (p=0.002 n=10+19)
BucketIndexReader_ExpandedPostings/i=~"^.+$",j=~"X.+"-16                      4.41ms ± 6%    4.38ms ± 9%     ~     (p=0.694 n=9+20)
BucketIndexReader_ExpandedPostings/i=~""-16                                    330ms ± 1%     336ms ± 2%   +1.98%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/i!=""-16                                    297ms ± 1%     298ms ± 1%     ~     (p=0.243 n=9+19)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-16                   29.7ms ± 1%    30.1ms ± 0%   +1.35%  (p=0.000 n=10+17)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",j="foo"-16                 2.97µs ±12%    2.86µs ±16%     ~     (p=0.169 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-16            33.6ms ± 1%    34.0ms ± 1%   +1.23%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!=""-16                              149ms ± 2%     149ms ± 1%   +0.51%  (p=0.044 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-16                      154ms ± 2%     156ms ± 1%   +1.08%  (p=0.001 n=10+16)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"X.+"-16                    3.90ms ±10%    3.72ms ± 9%   -4.57%  (p=0.031 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"XXX|YYY"-16                3.37µs ± 9%    3.49µs ± 9%     ~     (p=0.071 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",i=~"X|Y|Z",j="foo"-16                3.82µs ± 7%    3.83µs ±11%     ~     (p=0.854 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-16                    154ms ± 1%     156ms ± 1%   +1.69%  (p=0.000 n=8+19)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-16                  26.0ms ± 2%    24.4ms ± 1%   -6.14%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-16             160ms ± 2%     163ms ± 3%   +1.79%  (p=0.005 n=8+19)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-16           182ms ± 2%     183ms ± 2%     ~     (p=0.064 n=10+18)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-16    3.26µs ± 8%    3.22µs ± 8%     ~     (p=0.388 n=10+18)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-16                     29.1µs ± 6%    28.7µs ± 1%     ~     (p=0.308 n=9+19)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-16                         29.6µs ± 6%    28.5µs ± 1%   -3.70%  (p=0.001 n=10+18)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-16                           28.9µs ± 3%    28.4µs ± 1%   -1.64%  (p=0.007 n=9+20)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-16                              124ms ± 2%     118ms ± 2%   -4.89%  (p=0.000 n=9+20)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-16                     129ms ± 4%     121ms ± 2%   -6.45%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/i=~"<unique_prefix>.+"-16                  2.66ms ± 7%    0.14ms ± 2%  -94.58%  (p=0.000 n=9+19)
BucketIndexReader_ExpandedPostings/n="1",i=~"<unique_prefix>.+"-16            4.19ms ±11%    2.25ms ±14%  -46.41%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",i!~"<unique_prefix>.+"-16            10.5ms ± 9%     9.1ms ±12%  -13.21%  (p=0.000 n=9+20)
BucketIndexReader_ExpandedPostings/p!=""-16                                   27.2ms ± 3%    26.9ms ± 6%     ~     (p=0.074 n=10+20)

name                                                                        old alloc/op   new alloc/op   delta
BucketIndexReader_ExpandedPostings/n="X"-16                                   1.01kB ± 0%    1.02kB ± 0%   +1.59%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="X",j="foo"-16                           1.34kB ± 0%    1.39kB ± 0%   +3.57%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="X",j!="foo"-16                          1.34kB ± 0%    1.39kB ± 0%   +3.57%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/j=~"XXX|YYY"-16                            1.02kB ± 0%    1.04kB ± 0%   +1.56%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/j=~"X.+"-16                                1.96kB ± 0%    0.95kB ± 0%  -51.43%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~"X|Y|Z"-16                              1.07kB ± 0%    1.09kB ± 0%   +1.49%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1"-16                                   12.7MB ± 0%    12.7MB ± 0%     ~     (p=0.787 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-16                           19.8MB ± 0%    19.8MB ± 0%     ~     (p=0.265 n=10+18)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-16                           19.8MB ± 0%    19.8MB ± 0%     ~     (p=0.422 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-16                          19.8MB ± 0%    19.8MB ± 0%     ~     (p=0.449 n=10+19)
BucketIndexReader_ExpandedPostings/i=~".*"-16                                  287MB ± 0%     287MB ± 0%     ~     (p=0.423 n=9+18)
BucketIndexReader_ExpandedPostings/i=~".+"-16                                  327MB ± 0%     327MB ± 0%     ~     (p=0.060 n=10+19)
BucketIndexReader_ExpandedPostings/i=~"^.+$",j=~"X.+"-16                      4.81MB ± 0%    4.81MB ± 0%   -0.02%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~""-16                                    113MB ± 0%     113MB ± 0%   +0.00%  (p=0.011 n=10+17)
BucketIndexReader_ExpandedPostings/i!=""-16                                    327MB ± 0%     327MB ± 0%   +0.00%  (p=0.002 n=9+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-16                   21.4MB ± 0%    21.4MB ± 0%   +0.00%  (p=0.000 n=8+17)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",j="foo"-16                 1.52kB ± 0%    1.58kB ± 0%   +4.21%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-16            22.7MB ± 0%    22.7MB ± 0%   +0.00%  (p=0.000 n=8+20)
BucketIndexReader_ExpandedPostings/n="1",i!=""-16                             86.4MB ± 0%    86.4MB ± 0%     ~     (p=0.087 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-16                     93.5MB ± 0%    93.5MB ± 0%   +0.00%  (p=0.000 n=10+18)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"X.+"-16                    4.81MB ± 0%    4.81MB ± 0%   -0.02%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"XXX|YYY"-16                1.55kB ± 0%    1.62kB ± 0%   +4.12%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~"X|Y|Z",j="foo"-16                1.65kB ± 0%    1.71kB ± 0%   +3.88%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-16                   93.5MB ± 0%    93.5MB ± 0%   +0.00%  (p=0.000 n=10+18)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-16                  23.9MB ± 0%    23.9MB ± 0%   +0.00%  (p=0.000 n=9+18)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-16            94.8MB ± 0%    94.8MB ± 0%     ~     (p=0.422 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-16           100MB ± 0%      99MB ± 0%   -0.18%  (p=0.000 n=9+20)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-16    1.70kB ± 0%    1.78kB ± 0%   +4.71%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-16                     13.9kB ± 0%    13.9kB ± 0%   +0.12%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-16                         13.8kB ± 0%    13.8kB ± 0%   +0.12%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-16                           13.8kB ± 0%    13.8kB ± 0%   +0.12%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-16                              285MB ± 0%     285MB ± 0%     ~     (p=0.088 n=10+20)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-16                     287MB ± 0%     287MB ± 0%   +0.00%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/i=~"<unique_prefix>.+"-16                  1.61MB ± 0%    0.90MB ± 0%  -44.25%  (p=0.000 n=10+19)
BucketIndexReader_ExpandedPostings/n="1",i=~"<unique_prefix>.+"-16            2.91MB ± 0%    2.19MB ± 0%  -24.53%  (p=0.000 n=8+20)
BucketIndexReader_ExpandedPostings/n="1",i!~"<unique_prefix>.+"-16            14.3MB ± 0%    13.6MB ± 0%   -4.98%  (p=0.000 n=10+18)
BucketIndexReader_ExpandedPostings/p!=""-16                                   58.8MB ± 0%    58.8MB ± 0%   +0.00%  (p=0.002 n=10+19)

name                                                                        old allocs/op  new allocs/op  delta
BucketIndexReader_ExpandedPostings/n="X"-16                                     23.0 ± 0%      23.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="X",j="foo"-16                             28.0 ± 0%      28.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="X",j!="foo"-16                            28.0 ± 0%      28.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/j=~"XXX|YYY"-16                              24.0 ± 0%      24.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/j=~"X.+"-16                                  23.0 ± 0%      22.0 ± 0%   -4.35%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~"X|Y|Z"-16                                25.0 ± 0%      25.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1"-16                                     93.0 ± 0%      93.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-16                              122 ± 1%       121 ± 0%     ~     (p=0.245 n=10+20)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-16                              122 ± 1%       122 ± 0%     ~     (p=0.500 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-16                             120 ± 1%       120 ± 1%     ~     (p=1.096 n=10+20)
BucketIndexReader_ExpandedPostings/i=~".*"-16                                    107 ± 0%       107 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~".+"-16                                   600k ± 0%      600k ± 0%     ~     (p=0.181 n=6+19)
BucketIndexReader_ExpandedPostings/i=~"^.+$",j=~"X.+"-16                        28.0 ± 0%      27.0 ± 0%   -3.57%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/i=~""-16                                     600k ± 0%      600k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i!=""-16                                     600k ± 0%      600k ± 0%   +0.00%  (p=0.001 n=9+14)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-16                      124 ± 0%       125 ± 0%     ~     (p=0.700 n=10+20)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",j="foo"-16                   30.0 ± 0%      30.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-16               150 ± 0%       150 ± 0%     ~     (p=1.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!=""-16                               600k ± 0%      600k ± 0%     ~     (p=1.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-16                       600k ± 0%      600k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"X.+"-16                      31.0 ± 0%      30.0 ± 0%   -3.23%  (p=0.000 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i!="",j=~"XXX|YYY"-16                  30.0 ± 0%      30.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~"X|Y|Z",j="foo"-16                  33.0 ± 0%      33.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-16                     600k ± 0%      600k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-16                   66.8k ± 0%     66.8k ± 0%     ~     (p=0.700 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-16              600k ± 0%      600k ± 0%     ~     (p=0.131 n=10+20)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-16            667k ± 0%      667k ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="X",i=~"^.+$",i!~"^.*2.*$",j="foo"-16      31.0 ± 0%      31.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-16                        118 ± 0%       118 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-16                            118 ± 0%       118 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-16                              118 ± 0%       118 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-16                                165 ± 0%       165 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-16                       170 ± 0%       170 ± 0%     ~     (p=0.099 n=9+20)
BucketIndexReader_ExpandedPostings/i=~"<unique_prefix>.+"-16                    66.0 ± 0%      66.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i=~"<unique_prefix>.+"-16              93.0 ± 0%      93.0 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/n="1",i!~"<unique_prefix>.+"-16               123 ± 0%       123 ± 0%     ~     (all equal)
BucketIndexReader_ExpandedPostings/p!=""-16                                     98.0 ± 0%      98.0 ± 0%     ~     (all equal)
```

#### Which issue(s) this PR fixes or relates to

None.

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`