@colega 2023-01-23
_Draft until https://github.com/grafana/mimir-prometheus/pull/402 is merged, but ready for a first round of reviews_

I would also like to check why some benchmark cases (like `i!=""`) became slower, and write some test cases.